Removed the generic InterceptingIterator and repalced it with specific implementation
